requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
dataclasses==0.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
numpy==2.4.2
pandas==3.0.1
psutil==7.2.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
Werkzeug==3.1.6
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
paramiko
passlib
email-validator
